backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.1
sqlalchemy>=2.0.30
mysql-connector-python>=8.4.0
python-dotenv>=1.0.1
jinja2>=3.1.4